Include request event id in response tags
From NIP47: "The response event SHOULD contain an e tag with the id of
the request event it is responding to."

Otherwise you can't determine which request this response is responding
to. This probably doesn't need to be in a tag, it could be in the json.
In Damus I already subscribe to all events from the pubkey, so this
unnecessarily leaks metadata. Maybe something to bring up on nip47?

Signed-off-by: William Casarin <jb55@jb55.com>